버그 쉊ㅇ 1. 스킬 관련 데이터 못읽는 문제 2. 몹 정렬이 제대로 안되는 문제 3. 공격 시 주사위가 사망하는 문제

+++ UnityProject/Assets/Scripts/Common/FGlobal.cs
@@ -15,7 +15,7 @@ public class FGlobal
 
     public static float DiceCriticalChange = 0.33f;
 
-    public static int CalcEffectValue(FObjectBase InObject, int InValue, int InValuePerLevel, int InValuePerBattleLevel)
+    public static float CalcEffectValue(FObjectBase InObject, float InValue, float InValuePerLevel, float InValuePerBattleLevel)
     {
         FStatController statController = InObject.FindController<FStatController>();
         if (statController == null)
